Added payment mapper

+++ payment_service/src/main/java/com/bexos/payment_service/dto/PaymentRequest.java
@@ -8,7 +8,7 @@ public record PaymentRequest(
         Long orderId,
         Long price,
         @NotNull(message = "User id is required")
-        Integer userId,
+        Long userId,
         @NotBlank(message = "Reference is required")
         String reference,
         @NotBlank(message = "Phone Number is required")
